Add player data caching module and per-player endpoint
